+++ app/src/main/java/com/yanbin/reactivestickynote/MainActivity.kt
@@ -8,13 +8,13 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import com.yanbin.reactivestickynote.account.AccountService
-import com.yanbin.reactivestickynote.domain.EditTextViewModel
+import com.yanbin.reactivestickynote.edittext.EditTextViewModel
 import com.yanbin.reactivestickynote.login.LoginScreen
 import com.yanbin.reactivestickynote.login.LoginViewModel
-import com.yanbin.reactivestickynote.ui.vm.EditorViewModel
-import com.yanbin.reactivestickynote.ui.screen.EditTextScreen
-import com.yanbin.reactivestickynote.ui.screen.StickyNoteEditorScreen
-import com.yanbin.reactivestickynote.ui.screen.Screen
+import com.yanbin.reactivestickynote.editor.vm.EditorViewModel
+import com.yanbin.reactivestickynote.ui.route.EditTextScreen
+import com.yanbin.reactivestickynote.ui.route.StickyNoteEditorScreen
+import com.yanbin.reactivestickynote.ui.route.Screen
 import com.yanbin.reactivestickynote.ui.theme.ReactiveStickyNoteTheme
 import org.koin.android.ext.android.inject
 import org.koin.androidx.viewmodel.ext.android.viewModel
