+++ app/src/main/java/com/yanbin/reactivestickynote/di/NoteModule.kt
@@ -2,14 +2,14 @@ package com.yanbin.reactivestickynote.di
 
 import com.yanbin.reactivestickynote.account.AccountService
 import com.yanbin.reactivestickynote.account.SPAccountService
-import com.yanbin.reactivestickynote.data.FirebaseNoteRepository
-import com.yanbin.reactivestickynote.data.NoteRepository
-import com.yanbin.reactivestickynote.ui.vm.ContextMenuViewModel
-import com.yanbin.reactivestickynote.domain.EditTextViewModel
-import com.yanbin.reactivestickynote.ui.vm.EditorViewModel
-import com.yanbin.reactivestickynote.domain.StickyNoteEditor
+import com.yanbin.reactivestickynote.editor.data.FirebaseNoteRepository
+import com.yanbin.reactivestickynote.editor.data.NoteRepository
+import com.yanbin.reactivestickynote.editor.vm.ContextMenuViewModel
+import com.yanbin.reactivestickynote.edittext.EditTextViewModel
+import com.yanbin.reactivestickynote.editor.vm.EditorViewModel
+import com.yanbin.reactivestickynote.editor.domain.StickyNoteEditor
 import com.yanbin.reactivestickynote.login.LoginViewModel
-import com.yanbin.reactivestickynote.ui.vm.StickyNoteViewModel
+import com.yanbin.reactivestickynote.editor.vm.StickyNoteViewModel
 import org.koin.android.ext.koin.androidContext
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
