+++ app/src/main/java/com/yanbin/reactivestickynote/editor/view/StickyNoteView.kt
@@ -21,12 +21,10 @@ import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.LocalViewModelStoreOwner
-import com.yanbin.reactivestickynote.ui.vm.EditorViewModel
-import com.yanbin.reactivestickynote.model.StickyNote
-import com.yanbin.reactivestickynote.model.Position
-import com.yanbin.reactivestickynote.ui.vm.StickyNoteViewModel
+import com.yanbin.reactivestickynote.editor.model.StickyNote
+import com.yanbin.reactivestickynote.editor.model.Position
+import com.yanbin.reactivestickynote.editor.vm.StickyNoteViewModel
 import org.koin.androidx.viewmodel.ext.android.viewModel
-import java.util.*
 
 @Composable
 fun StatefulStickyNoteView(
